feat: moved everything in app into mainview (could rename to home). app now only contains routes to mainview and to item. will add route to addform in here as well
